more renaming, missing: SwapRZ
