+++ qiskit/circuit/library/n_local/ry_ansatz.py
@@ -22,7 +22,7 @@ from qiskit.extensions.standard import RYGate, CZGate
 from .two_local import TwoLocal
 
 
-class RY(TwoLocal):
+class RYAnsatz(TwoLocal):
     r"""The RY 2-local circuit.
 
     The RY trial wave function consistes of alternating layers of :math:`Y` rotations
@@ -64,7 +64,7 @@ class RY(TwoLocal):
 
     Examples:
 
-         >>> ry = RY(3, reps=2)  # create the variational form on 3 qubits
+         >>> ry = RYAnsatz(3, reps=2)  # create the variational form on 3 qubits
          >>> print(ry)  # show the circuit
               ┌──────────┐      ┌──────────┐                  ┌──────────┐
          q_0: ┤ RY(θ[0]) ├─■──■─┤ RY(θ[3]) ├─────────────■──■─┤ RY(θ[6]) ├────────────
@@ -74,7 +74,7 @@ class RY(TwoLocal):
          q_2: ┤ RY(θ[2]) ├────■──────■──────┤ RY(θ[5]) ├────■──────■──────┤ RY(θ[8]) ├
               └──────────┘                  └──────────┘                  └──────────┘
 
-         >>> ry = RY(3, entanglement='linear', reps=2, insert_barriers=True)
+         >>> ry = RYAnsatz(3, entanglement='linear', reps=2, insert_barriers=True)
          >>> qc = QuantumCircuit(3)  # create a circuit and append the RY variational form
          >>> qc += ry.to_circuit()
          >>> qc.decompose().draw()
@@ -86,7 +86,7 @@ class RY(TwoLocal):
          q_2: ┤ RY(θ[2]) ├─░─────■──░─┤ RY(θ[5]) ├─░─────■──░─┤ RY(θ[8]) ├
               └──────────┘ ░        ░ └──────────┘ ░        ░ └──────────┘
 
-         >>> ry = RY(4, 'crx', entanglement='circular', reps=2, insert_barriers=True)
+         >>> ry = RYAnsatz(4, 'crx', entanglement='circular', reps=2, insert_barriers=True)
          >>> print(ry)
               ┌──────────┐ ░ ┌──────────┐                                     ░  ┌──────────┐
          q_0: ┤ RY(θ[0]) ├─░─┤ Rx(θ[4]) ├─────■───────────────────────────────░──┤ RY(θ[8]) ├
@@ -99,7 +99,7 @@ class RY(TwoLocal):
               └──────────┘ ░                                     └──────────┘ ░ └───────────┘
 
          >>> entanglement = [[0, 1], [0, 2]]
-         >>> ry = RY(3, 'cx', entanglement, reps=2)
+         >>> ry = RYAnsatz(3, 'cx', entanglement, reps=2)
          >>> print(ry)
               ┌──────────┐                      ┌──────────┐                      ┌──────────┐
          q_0: ┤ RY(θ[0]) ├──■────────────────■──┤ RY(θ[3]) ├──■────────────────■──┤ RY(θ[6]) ├
